adds Page class to controllers, mv page specific functions
